BottomBar for import. Needs work, but need to look at whole body behaviour together. (currently scroll only takes place on inner container, which looks weird. i think we need a helper css to get limited-width for wide windows to avoid the nested containers!)
