+++ frontend/src/components/import/targetSelector.tsx
@@ -1,4 +1,5 @@
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
+
+import { Dispatch, SetStateAction, useState } from "react";
 import { Box, Button, FormHelperText } from "@mui/material";
 import Tooltip from "@mui/material/Tooltip";
 
@@ -120,46 +121,5 @@ function StartAndAbortBtn({
                 </Button>
             </Tooltip>
         );
-    }
-}
-
-export function ApplySelection() {
-    const { completeAllSelections, selectionsInvalidCause } = useImportContext();
-    const [tooltipText, setTooltipText] = useState<string | null>(null);
-
-    useEffect(() => {
-        if (selectionsInvalidCause === "no current candidate") {
-            setTooltipText("Pick a candidate to apply changes!");
-        } else if (selectionsInvalidCause === "no duplicate action") {
-            setTooltipText("Choose what to do with the duplciate candidates!");
-        } else {
-            setTooltipText(selectionsInvalidCause);
-        }
-    }, [selectionsInvalidCause]);
-
-    if (selectionsInvalidCause === "no selections") {
-        // hide the apply button when the session has not started yet
-        return null;
-    }
-
-    const inner = (
-        // Wrap into a Box to enable Tooltips on disabled buttons
-        <Box>
-            <Button
-                variant="outlined"
-                color="primary"
-                onClick={completeAllSelections}
-                disabled={selectionsInvalidCause !== null}
-                sx={{ width: "100%" }}
-            >
-                Apply
-            </Button>
-        </Box>
-    );
-
-    if (tooltipText) {
-        return <Tooltip title={tooltipText}>{inner}</Tooltip>;
-    } else {
-        return inner;
     }
 }
