+++ frontend/src/main.css
@@ -31,12 +31,15 @@
 
 /* applied to all custom hover cards that use radix */
 .HoverContent {
-    Background-color: #21252933;
     border-radius: 0.3rem;
-    backdrop-filter: blur(17px);
     padding: 1rem;
     z-index: 2;
     pointer-events: none;
+}
+
+.DefaultBlurBg, .HoverContent {
+    Background-color: #21252933;
+    backdrop-filter: blur(17px);
 }
 
 /* Ansi colors, opiniated, for https: //github.com/nteract/ansi-to-react */
